requirements.txt
aiohttp
asyncio
discord.py
pytz
psutil
schedule